backend/requirements.txt
anyio==4.12.1
beautifulsoup4==4.14.3
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.1.8
cryptography==46.0.5
defusedxml==0.7.1
exceptiongroup==1.3.1
Flask==3.1.3
flask-cors==6.0.2
google-api-core==2.30.0
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.3.0
googleapis-common-protos==1.73.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.2
MarkupSafe==3.0.3
notion-client==2.0.0
oauthlib==3.3.1
proto-plus==1.27.1
protobuf==6.33.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==2.23
PyMuPDF==1.26.5
pyparsing==3.3.2
python-dotenv==1.2.1
requests==2.32.5
requests-oauthlib==2.0.0
soupsieve==2.8.3
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
Werkzeug==3.1.7
youtube-transcript-api==1.2.4
zipp==3.23.0
